Web/requirement.txt
astroid==2.3.3
autopep8==1.4.4
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cssselect==1.1.0
decorator==4.4.1
dicttoxml==1.7.4
Django==2.2.3
django-crontab==0.7.1
django-ipware==2.1.0
django-qr-code==1.0.0
docutils==0.15.2
future==0.18.2
geocoder==1.38.1
geoip2==2.9.0
idna==2.8
ip2geotools==0.1.5
IP2Location==8.4.1
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.4.1
maxminddb==1.5.1
mccabe==0.6.1
packaging==19.2
pip-review==1.0
pkg-resources==0.0.0
pkginfo==1.5.0.1
pycodestyle==2.5.0
Pygments==2.4.2
pylint==2.4.4
pyparsing==2.4.5
pyquery==1.4.1
pytz==2019.1
qrcode==6.1
ratelim==0.1.6
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
selenium==3.141.0
send-email-message==0.1.5
six==1.13.0
sqlparse==0.3.0
tqdm==4.38.0
twine==2.0.0
typed-ast==1.4.0
typing==3.7.4.1
urllib3==1.25.7
webencodings==0.5.1
wrapt==1.11.2
